Add registration endpoint and benchmark routes
The registration blueprint contains the reverse benchmark endpoints, and
the registration websocket.

The main server url defaults to the public-facing https port (443),
which is routed to the internal application port.

The temporary server is bound to localhost for security.

These imports must be *after* forking to prevent event loop conflicts.